fontify most method signatures:
  * Add faces to Type $varname
  * Delete faces if signature does not parse
  * Support Type|Unions
  * Support "does coerce"
  * Support "does <foo>" and "is <foo>" (unused currently)
  * Support "Type $foo where { BLOCK }"
  * Support "Type $foo = EXPR"
  * Support multiple does/where/= clauses (not all make sense, but oh well)

Still broken are signatures with parens inside them.  Will fix later,
but docs to this effect are now here.

(Note also that there is some syntax-highlighting magic that hides
some bad effects of parens in the signature.  This is a side-effect of
syntax-highlighting the sub name as though it were a "use Foo"
declaration.  Not sure why I did this in the first place, but it is a
happy coincidence that it solves our problem.)