Fix indentation for qw().
This is a quick fix for issue #53. I have added a customizable user
option value "cperl-indentable-indent" which is by default set to "t"
which gives the previous default behavior. Only if the option is set
to "nil" by the user, the modification provided by the current pull
request is triggered.